update aio-lib-events to fetch provider metadata
