+++ lib/events/EventsOfInterestHelper.js
@@ -9,10 +9,18 @@ OF ANY KIND, either express or implied. See the License for the specific languag
 governing permissions and limitations under the License.
 */
 
+
+
+const {selectEventMetadataForProvider, selectProviderForProviderMetadata} = require("./ProviderHelper")
+const {getEntitledProviderMetadataForOrg, getProviderMetadata} = require("./ProviderMetadataHelper")
+
 async function promptForEventsOfInterest (eventsClient, obj) {
+  const providerMetadataList = await getEntitledProviderMetadataForOrg(eventsClient)
+  // get all providers filtered by the selected provider metadata ids
+  const providerMetadataIds = await getProviderMetadata(obj, providerMetadataList)
   // get a map of provider_metadata_id -> providers
-  const providerMetadataToProvidersMap = await getProviderMetadataToProvidersMap(obj.projectConfig.org.id, eventsClient)
-  const providerMetadataIds = await getProviderMetadata(obj, providerMetadataToProvidersMap)
+  const providerMetadataToProvidersMap = await getProviderMetadataToProvidersMap(obj.projectConfig.org.id, eventsClient, providerMetadataIds)
+
   // get the existing map of provider_metadata_id -> provider_id from the .env file
   const selectedProvidersToEventMetadata = new Map()
   for (const providerMetadata of providerMetadataIds) {
@@ -30,10 +38,14 @@ async function promptForEventsOfInterest (eventsClient, obj) {
 }
 
 /** @private */
-async function getProviderMetadataToProvidersMap (consumerId, eventsClient) {
-  const providersHalModel = await eventsClient.getAllProviders(consumerId, true)
+async function getProviderMetadataToProvidersMap (consumerId, eventsClient, providerMetadataIds) {
+  console.log(' Provider metadata ids selected : ', providerMetadataIds)
+  const providersHalModel = await eventsClient.getAllProviders(consumerId, { fetchEventMetadata: true,
+    filterBy: {
+      providerMetadataIds: providerMetadataIds
+    }
+  })
   const providers = providersHalModel._embedded.providers
-
   const providerMetadataToProvidersMap = providers.reduce((providerMetadataToProvidersMap, provider) => {
     providerMetadataToProvidersMap[provider.provider_metadata] = providerMetadataToProvidersMap[provider.provider_metadata] || []
     providerMetadataToProvidersMap[provider.provider_metadata].push({
@@ -60,80 +72,6 @@ function getProviderMetadataToProvidersExistingMap(providerMetadataToProviderMap
     return providerMetadataToProviderIdMappingExistingMap
   }
   return {}
-}
-
-/** @private */
-async function getProviderMetadata (obj, providerMetadataToProvidersMap) {
-  const { providerMetadataIds } = await obj.prompt({
-    type: 'checkbox',
-    name: 'providerMetadataIds',
-    message: 'Choose from below family of providers ( provider metadata )',
-    choices: Object.keys(providerMetadataToProvidersMap),
-    pageSize: 5,
-    validate: (options) => {
-      if (!options.length) {
-        return 'Choose at least one of the above, use space to choose the option'
-      }
-      return true
-    }
-  })
-  return providerMetadataIds
-}
-
-/** @private */
-async function selectProviderForProviderMetadata (providerMetadataToProvidersMap,
-    providerMetadata, obj) {
-  const providerChoices = []
-  const providersList = providerMetadataToProvidersMap[providerMetadata]
-  if (providersList.length === 1) {
-    providerChoices.push({
-      name: providersList[0].label,
-      description: providersList[0].description,
-      checked: true,
-      value: providersList[0]
-    })
-  } else {
-    for (const provider of providersList) {
-      providerChoices.push({
-        name: provider.label,
-        description: provider.description,
-        value: provider
-      })
-    }
-  }
-  const { provider } = await obj.prompt({
-    type: 'list',
-    name: 'provider',
-    message: `Choose from below provider for provider metadata: ${providerMetadata}`,
-    choices: providerChoices,
-    pageSize: 5
-  })
-  return provider
-}
-
-/** @private */
-async function selectEventMetadataForProvider (providerSelected, eventMetadataChoices, obj) {
-  for (const eventMetadata of providerSelected.eventMetadata) {
-    eventMetadataChoices.push({
-      name: eventMetadata.label,
-      description: eventMetadata.description,
-      value: eventMetadata.event_code
-    })
-  }
-  const { eventMetadataSelection } = await obj.prompt({
-    type: 'checkbox',
-    name: 'eventMetadataSelection',
-    message: `Choose event metadata for provider: ${providerSelected.label}`,
-    choices: eventMetadataChoices,
-    pageSize: 5,
-    validate: (options) => {
-      if (!options.length) {
-        return 'Choose at least one of the above, use space to choose the option'
-      }
-      return true
-    }
-  })
-  return eventMetadataSelection
 }
 
 module.exports = {
