+++ lib/events/RuntimeActionForEventsHelper.js
@@ -38,7 +38,7 @@ async function promptForRuntimeAction(obj) {
   }
   if (!runtimeActionName) {
     runtimeActionName = await obj.promptForActionName(
-        'showcases how to respond to IO Events', 'generic')
+        'A generic action that logs the events received from IO Events', 'generic')
   }
   return runtimeActionName;
 }
