+++ package.json
@@ -33,7 +33,7 @@
   "dependencies": {
     "@adobe/aio-lib-core-config": "^3.1.0",
     "@adobe/aio-lib-env": "^2.0.0",
-    "@adobe/aio-lib-events": "^3.0.0",
+    "@adobe/aio-lib-events": "^3.1.0",
     "@adobe/aio-lib-ims": "^6.3.0",
     "fs-extra": "^11.1.1",
     "js-yaml": "^4.1.0",
